+++ app/src/main/java/com/fibelatti/photowidget/preferences/UserPreferences.kt
@@ -1,5 +1,6 @@
 package com.fibelatti.photowidget.preferences
 
+import com.fibelatti.photowidget.model.DirectorySorting
 import com.fibelatti.photowidget.model.PhotoWidgetCycleMode
 import com.fibelatti.photowidget.model.PhotoWidgetSource
 import com.fibelatti.photowidget.model.PhotoWidgetTapAction
@@ -11,6 +12,7 @@ data class UserPreferences(
     val dynamicColors: Boolean,
     val defaultSource: PhotoWidgetSource,
     val defaultShuffle: Boolean,
+    val defaultDirectorySorting: DirectorySorting,
     val defaultCycleMode: PhotoWidgetCycleMode,
     val defaultShape: String,
     val defaultCornerRadius: Int,