+++ app/src/main/java/com/fibelatti/photowidget/preferences/WidgetDefaultsViewModel.kt
@@ -1,6 +1,7 @@
 package com.fibelatti.photowidget.preferences
 
 import androidx.lifecycle.ViewModel
+import com.fibelatti.photowidget.model.DirectorySorting
 import com.fibelatti.photowidget.model.PhotoWidgetCycleMode
 import com.fibelatti.photowidget.model.PhotoWidgetSource
 import com.fibelatti.photowidget.model.PhotoWidgetTapAction
@@ -21,6 +22,10 @@ class WidgetDefaultsViewModel @Inject constructor(
 
     fun saveDefaultShuffle(value: Boolean) {
         userPreferencesStorage.defaultShuffle = value
+    }
+
+    fun saveDefaultSorting(value: DirectorySorting) {
+        userPreferencesStorage.defaultDirectorySorting = value
     }
 
     fun saveDefaultCycleMode(cycleMode: PhotoWidgetCycleMode) {